Create services-ang.htm
Factory: A factory is a simple function which allows you to add some logic before creating the object and then it returns the created object.

Service: A service is a constructor function which creates the object using the new keyword. You can add properties and functions to a service object by using the this keyword. Unlike a factory, it doesn't return anything (it returns an object which contains method).

Provider: 
	

Understanding AngularJS Factory, Service and Provider

All of these are used to share reusable singleton objects, which are organize and also helps to share code across your app.

Factory

A factory is a simple function which allows you to add some logic before creating the object and then it returns the created object.

app.factory('MyFactory', function () {

 var serviceObj = {};
 serviceObj.myFunction = function () {
 //TO DO:
 };

 return serviceObj;
});

Usage

It is just a collection of functions like a class. Hence, it can be instantiated in different controllers when you are using it with constructor function.

Service

A service is a constructor function which creates the object using the new keyword. You can add properties and functions to a service object by using the this keyword. Unlike a factory, it doesn't return anything (it returns an object which contains method).

app.service('MyService', function () {
  this.function1 = function () {
  //TO DO:
  };
 });

Usage

Use it when you need to share a single object across the application. For example, authenticated user details, share-able data, etc.

Provider : A provider is used to create a configurable service object. It returns value by using $get() function. $get function gets executed on run phase angular.